requirements.txt
easyocr==1.7.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
spacy==3.7.2
pytesseract==0.3.10
requests==2.31.0
torch==2.1.0
torchvision==0.16.0
pytest==7.4.3
python-dotenv==1.0.0
transformers==4.38.1
accelerate==0.27.2
bitsandbytes==0.41.1
torch==2.1.0
einops==0.7.0